To Do Entities:
- Implement hasCollided for platforms.
- Complete all entities and add necessary methods.
- Player entity appears incomplete; unsure how to update the player as it requires user input. Might address this after the controller is finished.
- Investigate how to translate the current x, y values and width to screen coordinates and determine movement behavior.

In short, a lot of work still needed.

+++ src/Logic/Entities/Entity.cpp
@@ -2,7 +2,7 @@
 
 // Constructor of the entity
 
-Entity::Entity(float x, float y) : x(x), y(y) {}
+Entity::Entity(int x, int y) : x(x), y(y) {}
 
 // Destructor of the entity
